+++ R/deprecated.R
@@ -1,0 +1,6 @@
+#' @describeIn ypr_tabulate_schedule Tabulate Schedule (Deprecated)
+#' @export
+ypr_schedule <- function(population) {
+  deprecate_warn("0.4.0", "ypr_schedule()", "ypr_tabulate_schedule()")
+  ypr_tabulate_schedule(population)
+}