+++ R/namespace.R
@@ -1,4 +1,4 @@
-#' @import ggplot2 chk yesno
+#' @import ggplot2 chk yesno lifecycle
 #' @importFrom purrr map map2 transpose
 #' @importFrom stats update setNames
 #' @importFrom graphics plot